change intensity values to double from int

+++ src/main/java/com/app/fitness/fitnesprogramapp/dtos/metrics/IntensityMetricDto.java
@@ -8,8 +8,8 @@ import lombok.NoArgsConstructor;
 public class IntensityMetricDto {
     private Long id;
 
-    private int minimumIntensity;
-    private int maximumIntensity;
+    private double minimumIntensity;
+    private double maximumIntensity;
     private boolean isRange;
     private String title;
     private String metricSymbol;
