+++ src/main/java/com/app/fitness/fitnesprogramapp/dtos/program/programhistory/DoneSetsHistoryDTO.java
@@ -17,7 +17,7 @@ public class DoneSetsHistoryDTO {
 
     private int volume;
 
-    private int intensity;
+    private double intensity;
 
     private Date date;
